backend/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
class-registry==2.1.2
coreapi==2.3.3
coreschema==0.0.4
cryptography==44.0.2
defusedxml==0.7.1
diff-match-patch==20241021
Django==5.2
django-cors-headers==4.7.0
django-filter==25.1
django-import-export==4.3.7
django-templated-mail==1.1.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
djoser==2.3.1
drf-extra-fields==3.7.0
et_xmlfile==2.0.0
filetype==1.2.0
filters==1.3.2
gunicorn==23.0.0
idna==3.10
isort==6.0.1
itypes==1.2.0
Jinja2==3.1.6
MarkupPy==1.18
MarkupSafe==3.0.2
numpy==2.2.5
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.5
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
sqlparse==0.5.3
tablib==3.8.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
xlrd==2.0.1
xlwt==1.3.0
